Move calendar in its own file
